Improve simplification of generic functions.
Until now, deriving simplified definitions by figuring out which constants
needed to be left opaque on an ad hoc basis. This made the simplification code
harder to reason about. This commit implements a different approach, where we
list all the constants that *should* be unfolded in a hint database called
deriving_compute.  I couldn't find a way of using the hint database as an
argument to cbv (as opposed to unfold), so I've added a compilation step that
produces an appropriate version of cbv by inspecting the hint database.